unskipping  all the ui integrations tests
