+++ polaris-ui/cypress/e2e/case-details/redactionLog.cy.ts
@@ -29,7 +29,7 @@ const redactionTypeLabels = [
   "Other",
 ];
 
-xdescribe("Redaction Log", () => {
+describe("Redaction Log", () => {
   describe("Feature Flag On Under Redaction", () => {
     it("Should show the redaction types select input along with the redaction button and show under redaction modal on clicking save redaction with correct redaction type summary in descending order of redaction types count", () => {
       cy.visit("/case-details/12AB1111111/13401?redactionLog=true");
